requirements.txt
amqp==5.0.2
bcrypt==3.2.0
bigquery==0.0.1
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colab==1.13.5
cryptography==3.2.1
cssselect==1.1.0
dacktool==0.0.7
dbstream==0.0.20
diazo==1.4.0
discovery==0.0.4
Django==1.7.11
django-haystack==3.0
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==1.3.0
etiquetando==0.1
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
html2text==2020.1.16
httplib2==0.18.1
idna==2.10
kombu==5.0.2
lxml==4.6.1
oauth2client==4.1.3
oauthlib==3.1.0
paramiko==2.7.2
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
PyNaCl==1.4.0
pyTelegramBotAPI==3.7.4
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
repoze.lru==0.7
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sshtunnel==0.1.5
stemming==1.0.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
